Update old footer links
fttoolkit.co.uk is a redirect to commercial.ft.com and as it's a
non-ft.com domain it's at higher risk of no longer being owned
by the FT. Let's just skip the redirect and link to an ft.com subdomain.

The markets.ft.com page is broken and gives a 404. Update the link to a
page which is working.