Replace My Account with myFT in right nav
Early 2024 we ran a test to improve the header entry points to myFT and myaccount.

As part of this test the locations of myFT and My Account were swapped

That test was successful so we want to make this a permenant change.

More details about that test in this PR https://github.com/Financial-Times/dotcom-page-kit/pull/1017