Add new nav bar right navigation objects
Previously there was 2 objects

1. navbar-right
2. navbar-right-anon

These sound like they are data for the same area of the header just based on different user states (anon vs logged in)

However they were being used in different placements within the header.

navbar-right = This is used in the second/bottom row of the header
navr-bar-right-non = This is used in the first/top row of the header

This causes confusion when trying to make changes to the navigation data and header.

To get around this we are adding a 2 new objects

1. navbar-top-right
2. navbar-top-right-anon

These should always in the top/first row of the header.

In order to keep backwards compatibility for the timebeing we are keeping navbar-right-anon but this is
an alias to the same data as navbar-top-right-anon.

This also enables us to roll out changes related to the AB test that the Retention team ran.

It allows us to keep the my account link within the navbar-top-right slot when logged in.